pyproject.toml
# pyproject.toml (minimal)
[project]
name = "spec2dv"
version = "0.1.0"
description = "Design Spec -> SQL/XML -> DV outputs (Spec2DV Data Modeler)"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "typer>=0.9.0",
]

[project.scripts]
spec2dv = "spec2dv.cli:app"